emit 'render' event for async nodes
When mounting a node, its parent may already be in the DOM. If so, the node is batched so any children can be added for free. You may want to know when an "async node" is finally rendered. If so, you can listen for the 'render' event on the node!

The 'render' event does *not* bubble up to ancestors.