fix(tween): camelcase keys
Camelcase keys (eg: backgroundColor) were converted to kebab and used to extract from the `to` and `from` config objects. We need to use the keys as-is when extracting from those objects.

This commit also fixes a incorrect reference, which resulted in the animation not being removed from the node's animation cache once completed.